+++ rq/cli/helpers.py
@@ -80,6 +80,8 @@ def get_redis_from_config(settings, connection_class=Redis):
         'ssl_ca_certs': settings.get('REDIS_SSL_CA_CERTS', None),
         'ssl_cert_reqs': settings.get('REDIS_SSL_CERT_REQS', 'required'),
         'ssl_ca_data': settings.get('REDIS_SSL_CA_DATA', None),
+        'ssl_keyfile': settings.get('REDIS_SSL_KEYFILE', None),
+        'ssl_certfile': settings.get('REDIS_SSL_CERTFILE', None),
     }
 
     return connection_class(**kwargs)