+++ rq/utils.py
@@ -48,7 +48,7 @@ def as_text(v: Union[bytes, str]) -> str:
         ValueError: If the value is not bytes or string
 
     Returns:
-        value (Optional[str]): Either the decoded string or None
+        value (str): The decoded string
     """
     if isinstance(v, bytes):
         return v.decode('utf-8')
@@ -124,10 +124,6 @@ def import_attribute(name: str) -> Callable[..., Any]:
     return getattr(attribute_owner, attribute_name)
 
 
-def utcnow():
-    return datetime.datetime.utcnow()
-
-
 def now():
     """Return now in UTC"""
     return datetime.datetime.now(datetime.timezone.utc)
@@ -145,7 +141,10 @@ def utcparse(string: str) -> dt.datetime:
         return datetime.datetime.strptime(string, _TIMESTAMP_FORMAT)
     except ValueError:
         # This catches any jobs remain with old datetime format
-        return datetime.datetime.strptime(string, '%Y-%m-%dT%H:%M:%SZ')
+        try:
+            return datetime.datetime.strptime(string, '%Y-%m-%dT%H:%M:%S.%fZ')
+        except ValueError:
+            return datetime.datetime.strptime(string, '%Y-%m-%dT%H:%M:%SZ')
 
 
 def first(iterable: Iterable, default=None, key=None):
@@ -224,7 +223,7 @@ def current_timestamp() -> int:
     Returns:
         int: _description_
     """
-    return calendar.timegm(datetime.datetime.utcnow().utctimetuple())
+    return calendar.timegm(datetime.datetime.now(datetime.timezone.utc).utctimetuple())
 
 
 def backend_class(holder, default_name, override=None) -> TypeVar('T'):